@@ -20,12 +20,12 @@
     "require": {
         "memio/model": "^3.0",
         "memio/pretty-printer": "^3.0",
-        "php": "^7.2 || ^8.0",
-        "twig/twig": "^1.18 || ^2.0 || ^3.0"
+        "php": "^8.0",
+        "twig/twig": "^3.0"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.19.3",
-        "phpspec/phpspec": "^6.1 || ^7.0",
+        "friendsofphp/php-cs-fixer": "^3.0",
+        "phpspec/phpspec": "^7.0",
         "rector/rector": "^2.3",
         "rector/swiss-knife": "^2.3",
         "phpstan/phpstan": "^2.1"
