@@ -13,7 +13,6 @@ namespace Memio\TwigTemplateEngine\TwigExtension;
 
 use Memio\Model\FullyQualifiedName;
 use Memio\Model\Phpdoc\ParameterTag;
-use Memio\Model\Type as ModelType;
 use Memio\TwigTemplateEngine\TwigExtension\Line\Line;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
@@ -21,11 +20,8 @@ use Twig\TwigFunction;
 
 class Whitespace extends AbstractExtension
 {
-    private $line;
-
-    public function __construct(Line $line)
+    public function __construct(private Line $line)
     {
-        $this->line = $line;
     }
 
     public function getFunctions(): array
@@ -48,10 +44,9 @@ class Whitespace extends AbstractExtension
         $elementLength = strlen($current);
         $longestElement = $elementLength;
         foreach ($collection as $element) {
-            if (ParameterTag::class === get_class($element)) {
-                $type = $element->getType();
-                $modelType = new ModelType($element->getType());
-                if ($modelType->isObject()) {
+            if (ParameterTag::class === $element::class) {
+                $type = $element->type->getName();
+                if ($element->type->isObject()) {
                     $fullyQualifiedName = new FullyQualifiedName($type);
                     $type = $fullyQualifiedName->getName();
                 }
@@ -65,7 +60,7 @@ class Whitespace extends AbstractExtension
     public function indent(
         string $text,
         int $level = 1,
-        string $type = 'code'
+        string $type = 'code',
     ): string {
         $lines = explode("\n", $text);
         $indentedLines = [];
